Recommendation

+++ app/src/main/java/com/example/moviepicker/domain/builder/RatingBuilder.kt
@@ -6,11 +6,11 @@ import com.example.moviepicker.domain.items.RatingItem
 class RatingBuilder {
     companion object {
         fun toDTO(ratingItem: RatingItem): RatingDTO {
-            return RatingDTO(ratingItem.userId, ratingItem.movieId, ratingItem.rating.toFloat())
+            return RatingDTO(ratingItem.userId, ratingItem.movieId, ratingItem.rating)
         }
 
         fun toItem(ratingDTO: RatingDTO): RatingItem {
-            return RatingItem(ratingDTO.userId, ratingDTO.movieId, ratingDTO.rating.toInt())
+            return RatingItem(ratingDTO.userId, ratingDTO.movieId, ratingDTO.rating)
         }
     }
 }
