+++ app/src/main/java/com/example/moviepicker/domain/items/RatingItem.kt
@@ -1,3 +1,3 @@
 package com.example.moviepicker.domain.items
 
-data class RatingItem(val userId: Int, val movieId: Int, val rating: Int)
+data class RatingItem(val userId: Int, val movieId: Int, val rating: Float)
